Annotating and formatting code. Questions for Tom
Tom, I have gone through the code for the two functions now and have heavily annotated it and reformatted it so as to make life easier for when I build a package from this. Where I have questions for you, I have written your name in capitals (TOM) so you can just Control+F to find the questions quickly.

Co-Authored-By: reedtom <25622282+reedtom@users.noreply.github.com>